delete ai units before next mission starts
currently only dead ai are cleaned up
this adds a flag to the spawn script to identify the mission type they
are associated with. once a mission starts, all the units associated
with the previous mission of that type are purged using epoch cleanup
code.

buildings/etc could also be added into this.